requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
moviepy==2.2.1
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
replicate==0.31.0
